+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-04-26 10:20:20
-// Commit: feat(core): implement shimmer loading skeleton in core
+// Updated at: 2026-04-26 11:28:47
+// Commit: fix(core): prevent double tap duplicate action in core
 
 class coreComponentState {
-  final String commitRef = '700';
-  final String lastMessage = 'feat(core): implement shimmer loading skeleton in core';
-  final DateTime timestamp = DateTime.parse('2026-04-26T10:20:20+05:00');
+  final String commitRef = '702';
+  final String lastMessage = 'fix(core): prevent double tap duplicate action in core';
+  final DateTime timestamp = DateTime.parse('2026-04-26T11:28:47+05:00');
 }