+++ CHANGELOG.md
@@ -703,3 +703,4 @@
 - [2026-04-26 10:07] feat(network): add bi-directional event bus handler for network
 - [2026-04-26 10:20] feat(core): implement shimmer loading skeleton in core
 - [2026-04-26 10:34] test(navigation): implement widget integration tests for navigation
+- [2026-04-26 11:28] fix(core): prevent double tap duplicate action in core